[refactor] rename learn → memory, observe → telemetry, evolve → calibration
Terms that don't collide with RL concepts:

  learn/    → memory/       (recording + recalling experience)
  observe/  → telemetry/    (read-only system visualization)
  evolve    → calibration   (tuning agents to your project)

Directory renames:
  learn/    → memory/     (20 files)
  observe/  → telemetry/  (4 files)

Skill renames:
  skills/learn/  → skills/memory/     (name: memory)
  skills/evolve/ → skills/calibration/ (name: calibration)

CLI:
  dynos calibration  (new name, evolve still works as alias)
  dynos memory       (new name)

All hooks/ wrappers updated. 909 passed, 0 failed.
