+++ bin/dynos
@@ -41,7 +41,8 @@ Internals (advanced):
   patterns      Generate patterns from retrospectives
   ctl           Task lifecycle, approvals, validation
   postmortem    Postmortems and improvements
-  evolve        Learned agent management
+  calibration   Agent calibration (generate, benchmark, promote/demote)
+  memory        Memory aggregation (patterns, trajectories)
   trajectory    Trajectory store operations
   bench         Benchmark runner
   report        Generate reports
@@ -260,7 +261,7 @@ print('no')
   patterns)   exec python3 "${HOOKS_DIR}/patterns.py" "$@" ;;
   ctl)        exec python3 "${HOOKS_DIR}/ctl.py" "$@" ;;
   postmortem) exec python3 "${HOOKS_DIR}/postmortem.py" "$@" ;;
-  evolve)     exec python3 "${HOOKS_DIR}/evolve.py" "$@" ;;
+  calibration|evolve) exec python3 "${HOOKS_DIR}/evolve.py" "$@" ;;
   trajectory) exec python3 "${HOOKS_DIR}/trajectory.py" "$@" ;;
   bench)      exec python3 "${HOOKS_DIR}/bench.py" "$@" ;;
   report)     exec python3 "${HOOKS_DIR}/report.py" "$@" ;;
