+++ hooks/auto.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
-"""Compatibility wrapper — implementation moved to learn/auto.py."""
+"""Compatibility wrapper — implementation moved to memory/auto.py."""
 import sys as _sys
 _sys.path.insert(0, str(__import__("pathlib").Path(__file__).resolve().parent.parent))
 _sys.path.insert(0, str(__import__("pathlib").Path(__file__).resolve().parent))
 _is_main = __name__ == "__main__"
-from learn import auto as _real
+from memory import auto as _real
 # Make the real module importable under the old name too
 _sys.modules[__name__ if not _is_main else "auto"] = _real
 if not _is_main:
