+++ hooks/generate.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
-"""Compatibility wrapper — implementation moved to learn/generate.py."""
+"""Compatibility wrapper — implementation moved to memory/generate.py."""
 import sys as _sys
 _sys.path.insert(0, str(__import__("pathlib").Path(__file__).resolve().parent.parent))
 _sys.path.insert(0, str(__import__("pathlib").Path(__file__).resolve().parent))
 _is_main = __name__ == "__main__"
-from learn import generate as _real
+from memory import generate as _real
 # Make the real module importable under the old name too
 _sys.modules[__name__ if not _is_main else "generate"] = _real
 if not _is_main:
