+++ memory/auto.py
