+++ memory/bench.py
