+++ memory/fixture.py
