+++ memory/generate.py
