+++ memory/lib_registry.py
