+++ memory/lib_trajectory.py
