+++ memory/patterns.py
