+++ memory/sweeper.py
