+++ memory/trajectory.py
