+++ skills/memory/contract.json
