+++ telemetry/global_stats.py
